requirements.txt
streamlit
sqlalchemy
psycopg2-binary
google-auth
google-auth-oauthlib
streamlit-authenticator